Rename module from entity/Tab to entity/EntitiesTab
The module should have the same name as the component itself.
