+++ src/web/pages/certbund/DetailsPage.jsx
@@ -29,9 +29,9 @@ import TableHeader from 'web/components/table/Header';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/TableRow';
 import DetailsBlock from 'web/entity/Block';
+import EntitiesTab from 'web/entity/EntitiesTab';
 import EntityComponent from 'web/entity/EntityComponent';
 import EntityPage from 'web/entity/EntityPage';
-import EntitiesTab from 'web/entity/Tab';
 import EntityTags from 'web/entity/Tags';
 import withEntityContainer from 'web/entity/withEntityContainer';
 import useTranslation from 'web/hooks/useTranslation';
