+++ src/web/pages/cpes/DetailsPage.jsx
@@ -31,10 +31,10 @@ import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/StripedTable';
 import TableRow from 'web/components/table/TableRow';
 import DetailsBlock from 'web/entity/Block';
+import EntitiesTab from 'web/entity/EntitiesTab';
 import EntityComponent from 'web/entity/EntityComponent';
 import {InfoLayout} from 'web/entity/EntityInfo';
 import EntityPage from 'web/entity/EntityPage';
-import EntitiesTab from 'web/entity/Tab';
 import EntityTags from 'web/entity/Tags';
 import withEntityContainer from 'web/entity/withEntityContainer';
 import useTranslation from 'web/hooks/useTranslation';
