+++ src/web/pages/reportformats/DetailsPage.jsx
@@ -28,13 +28,13 @@ import TableHead from 'web/components/table/Head';
 import TableHeader from 'web/components/table/Header';
 import Table from 'web/components/table/StripedTable';
 import TableRow from 'web/components/table/TableRow';
+import EntitiesTab from 'web/entity/EntitiesTab';
 import EntityPage from 'web/entity/EntityPage';
 import CreateIcon from 'web/entity/icon/CreateIcon';
 import EditIcon from 'web/entity/icon/EditIcon';
 import TrashIcon from 'web/entity/icon/TrashIcon';
 import {goToDetails, goToList} from 'web/entity/navigation';
 import EntityPermissions from 'web/entity/Permissions';
-import EntitiesTab from 'web/entity/Tab';
 import EntityTags from 'web/entity/Tags';
 import withEntityContainer, {
   permissionsResourceFilter,
