+++ src/web/pages/results/DetailsPage.jsx
@@ -43,11 +43,11 @@ import TableData from 'web/components/table/Data';
 import InfoTable from 'web/components/table/InfoTable';
 import TableRow from 'web/components/table/TableRow';
 import DetailsBlock from 'web/entity/Block';
+import EntitiesTab from 'web/entity/EntitiesTab';
 import EntityPage from 'web/entity/EntityPage';
 import {goToDetails} from 'web/entity/navigation';
 import Note from 'web/entity/Note';
 import Override from 'web/entity/Override';
-import EntitiesTab from 'web/entity/Tab';
 import EntityTags from 'web/entity/Tags';
 import withEntityContainer from 'web/entity/withEntityContainer';
 import useCapabilities from 'web/hooks/useCapabilities';
